+++ app/api/api.py
@@ -5,36 +5,18 @@ import io
 import logging
 import os
 import sys
-from llm_reviewer.agent import *
-from llm_reviewer.brevity import *
-from llm_reviewer.style import *
-from llm_reviewer.impact import *
-from llm_reviewer.jd import structure_jd,calculate_score_llm
-from utils.utils import *
-from llm_reviewer.summary_prompt import master_prompt
-import uvicorn
-from fastapi.middleware.cors import CORSMiddleware
-
-app = FastAPI()
-
-origins = [
-    "http://localhost",
-    "http://localhost:8080",
-    "http://localhost:3000"
-]
-
-app.add_middleware(
-    CORSMiddleware,
-    allow_origins=origins,
-    allow_credentials=True,
-    allow_methods=["*"],
-    allow_headers=["*"],
-)
-# Initialize FastAPI app
+from app.llm_reviewer.agent import *
+from app.llm_reviewer.brevity import *
+from app.llm_reviewer.style import *
+from app.llm_reviewer.impact import *
+from app.utils.utils import *
 
 # Setup logging
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
+
+# Initialize FastAPI app
+app = FastAPI()
 
 # Root endpoint
 @app.get("/")
@@ -52,36 +34,13 @@ class TextList(BaseModel):
 class TextInput(BaseModel):
     cv_text: str
 
-class TextInputJob(BaseModel):
-    job_text: str
-
 class DictInput(BaseModel):
-    extracted_data: Dict[str, Any]
+    data: Dict[str, Any]
 
 class SkillCheckerRequest(BaseModel):
     hard_skills: List[str]
     soft_skills: List[str]
     profile: str = "Full Stack Web Development"
-
-import re
-
-def clean_text(text):
-    # Remove single quotes
-    text = re.sub(r"'", "", text)
-    
-    # Remove double quotes
-    text = re.sub(r'"', '', text)
-    
-    return text
-
-@app.post("/summary")
-async def summary_resume(cv_text:TextInput):
-    try:
-        return {"message": master_prompt(clean_text(cv_text.cv_text))}
-    except Exception as e:
-        logger.error(f"Error extracting Summary data: {e}")
-        raise HTTPException(status_code=500, detail="Error extracting Summary data")
-
 
 # Endpoint to extract text from PDF
 @app.post("/extract_text_from_pdf")
@@ -95,34 +54,12 @@ async def extract_pdf_text(file: UploadFile = File(...)):
         logger.error(f"Error extracting text from PDF: {e}")
         raise HTTPException(status_code=500, detail="Error extracting text from PDF")
 
-@app.post("/structure_job_description")
-async def extract_job_description(input : TextInputJob):
-    try:
-        return {"message": structure_jd(input.job_text)}
-    except Exception as e:
-        logger.error(f"Error extracting structured data: {e}")
-        raise HTTPException(status_code=500, detail="Error extracting structured data")
-    
-@app.post("/job_description_resume_score")
-async def calculate_score(cv_text:TextInput, job_text : TextInputJob):
-    try:
-        scores = calculate_score_llm(clean_text(cv_text.cv_text),job_text.job_text)
-        hard_skills_score = scores["HARD_SKILLS_SCORE"]['score']
-        soft_skills_score = scores["SOFT_SKILLS_SCORE"]['score']
-        experience_score = scores["EXPERIENCE_SCORE"]['score']
-        education_score = scores["EDUCATION_SCORE"]['score']
-        final_score = int((((0.4 * hard_skills_score) + (0.2 * experience_score) + (0.2 * education_score) + (0.2 * soft_skills_score))/175)*100)
-        return {"message": {"FINAL_SCORE":final_score,"DETAILS":scores}}
-    except Exception as e:
-        logger.error(f"Error extracting structured data: {e}")
-        raise HTTPException(status_code=500, detail="Error extracting structured data")
- 
 
 # Endpoint to extract structured data from text
 @app.post("/extract_structured_data")
 async def extract_data(input: TextInput):
     try:
-        return {"message": extract_structured_data(clean_text(input.cv_text))}
+        return {"message": extract_structured_data(input.cv_text)}
     except Exception as e:
         logger.error(f"Error extracting structured data: {e}")
         raise HTTPException(status_code=500, detail="Error extracting structured data")
@@ -131,7 +68,7 @@ async def extract_data(input: TextInput):
 @app.post("/quantification")
 async def analyze_quantification(data:DictInput):
     try:
-        input = "\n".join(data.extracted_data["Description"])
+        input = "\n".join(data["Description"])
         return {"message": quantification(input)}
     except Exception as e:
         logger.error(f"Error analyzing quantification: {e}")
@@ -139,20 +76,19 @@ async def analyze_quantification(data:DictInput):
 
 # Endpoint for repetition analysis
 @app.post("/repetition")
-async def analyze_repetition(data: DictInput):
+async def analyze_repetition(data):
     try:
-        input = "\n".join(data.extracted_data["Description"])
-        length = len(data.extracted_data["Description"])
-        return {"message": repetition(input,length)}
+        input = "\n".join(data["Description"])
+        return {"message": repetition(input)}
     except Exception as e:
         logger.error(f"Error analyzing repetition: {e}")
         raise HTTPException(status_code=500, detail="Error analyzing repetition")
 
 # Endpoint for weak verb checking
 @app.post("/weak_verb_checker")
-async def check_weak_verbs(data: DictInput):
+async def check_weak_verbs(data):
     try:
-        input = "\n".join(data.extracted_data["Description"])
+        input = "\n".join(data["Description"])
         return {"message": weak_verb_checker(input)}
     except Exception as e:
         logger.error(f"Error checking weak verbs: {e}")
@@ -160,9 +96,9 @@ async def check_weak_verbs(data: DictInput):
 
 # Endpoint for verb tense checking
 @app.post("/verb_tense")
-async def check_verb_tense(data: DictInput):
+async def check_verb_tense(data):
     try:
-        input = "\n".join(data.extracted_data["Description"])
+        input = "\n".join(data["Description"])
         return {"message": verb_tense(input)}
     except Exception as e:
         logger.error(f"Error checking verb tense: {e}")
@@ -170,9 +106,9 @@ async def check_verb_tense(data: DictInput):
 
 # Endpoint for responsibility analysis
 @app.post("/responsibility")
-async def analyze_responsibility(data: DictInput):
+async def analyze_responsibility(data):
     try:
-        input = "\n".join(data.extracted_data["Description"])
+        input = "\n".join(data["Description"])
         return {"message": responsibility(input)}
     except Exception as e:
         logger.error(f"Error analyzing responsibility: {e}")
@@ -180,9 +116,9 @@ async def analyze_responsibility(data: DictInput):
 
 # Endpoint for spelling checking
 @app.post("/spelling_checker")
-async def check_spelling(data: DictInput):
+async def check_spelling(data):
     try:
-        input = "\n".join(data.extracted_data["Description"])
+        input = "\n".join(data["Description"])
         return {"message": spelling_checker(input)}
     except Exception as e:
         logger.error(f"Error checking spelling: {e}")
@@ -199,29 +135,27 @@ async def analyze_resume_length(request: ResumeAnalysisRequest):
 
 # Endpoint for bullet point length analysis
 @app.post("/bullet_point_length")
-async def analyze_bullet_point_length(data: DictInput):
+async def analyze_bullet_point_length(data):
     try:
-        input = "\n".join(data.extracted_data["Description"])
-        return {"message": bullet_point_length(input)}
+        return {"message": bullet_point_length(input.points_to_check)}
     except Exception as e:
         logger.error(f"Error analyzing bullet point length: {e}")
         raise HTTPException(status_code=500, detail="Error analyzing bullet point length")
 
 # Endpoint for total bullet list analysis
 @app.post("/total_bullet_list")
-async def analyze_total_bullet_list(data:DictInput):
+async def analyze_total_bullet_list(cv_data,experience):
     try:
-        input = "\n".join(data.extracted_data["Description"])
-        return {"message": total_bullet_list(input)}
+        return {"message": total_bullet_list(cv_data,experience)}
     except Exception as e:
         logger.error(f"Error analyzing total bullet list: {e}")
         raise HTTPException(status_code=500, detail="Error analyzing total bullet list")
 
 # Endpoint to improve bullet points
 @app.post("/bullet_points_improver")
-async def improve_bullet_points(data: DictInput):
+async def improve_bullet_points(data):
     try:
-        input = "\n".join(data.extracted_data["Description"])
+        input = "\n".join(data["Description"])
         return {"message": bullet_points_improver(input)}
     except Exception as e:
         logger.error(f"Error improving bullet points: {e}")
@@ -229,9 +163,9 @@ async def improve_bullet_points(data: DictInput):
 
 # Endpoint for personal info checking
 @app.post("/personal_info")
-async def check_personal_info(data: DictInput):
+async def check_personal_info(data):
     try:
-        input = "\n".join(data.extracted_data["Personal Information"])
+        input = "\n".join(data["Personal Information"])
         return {"message": personal_info(input)}
     except Exception as e:
         logger.error(f"Error checking personal info: {e}")
@@ -239,9 +173,9 @@ async def check_personal_info(data: DictInput):
 
 # Endpoint for section checking
 @app.post("/section_checker")
-async def check_sections(data: DictInput):
+async def check_sections(data):
     try:
-        input = "\n".join(data.extracted_data["Sections"])
+        input = "\n".join(data["Sections"])
         return {"message": section_checker(input)}
     except Exception as e:
         logger.error(f"Error checking sections: {e}")
@@ -249,15 +183,12 @@ async def check_sections(data: DictInput):
 
 # Endpoint for skill checking
 @app.post("/skill_checker")
-async def check_skills(data: DictInput, profile :str):
+async def check_skills(data, profile):
     try:
-        hard_skills = data.extracted_data["Skills"]["HARD"]
-        soft_skills = data.extracted_data["Skills"]["SOFT"]
+        hard_skills = data["Skills"]["HARD"]
+        soft_skills = data["Skills"]["SOFT"]
         profile = profile
         return {"message": skill_checker(hard_skills, soft_skills, profile)}
     except Exception as e:
         logger.error(f"Error checking skills: {e}")
         raise HTTPException(status_code=500, detail="Error checking skills")
-
-if __name__ == "__main__":
-    uvicorn.run(app)
