+++ app/utils/utils.py
@@ -1,6 +1,6 @@
 import streamlit as st
 import PyPDF2
-from llm_reviewer.agent import *
+from app.llm_reviewer.agent import *
 import json
 import sys
 import os
